+++ CHANGELOG.md
@@ -12,6 +12,8 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 * Register `pathlib.Path` with `ub.repr2`
 * Can now register global `ub.repr2` extensions `ub.repr2.register`
 * Can now register global `ub.hash_data` extensions `ub.hash_data..register`
+* `ub.cmd` now has a `system` argument for modularity with `os.system`.
+* Removed deprecated arguments from `ubelt.cmd`.
 
 
 ## Version 1.0.1 - Released 2022-02-20